+++ src/Modules/Store/PickupValidationException.php
@@ -2,19 +2,18 @@
 
 namespace Foodsharing\Modules\Store;
 
-use Exception;
-use Throwable;
-
-class PickupValidationException extends Exception
+class PickupValidationException extends \InvalidArgumentException
 {
+	public const MAX_SLOT_COUNT_OUT_OF_RANGE = 'Slot value out of range.';
+	public const DUPLICATE_PICKUP_DAY_TIME = 'Multiply pickups for the same day and time.';
 	public const PICK_UP_DATE_IN_THE_PAST = 'Pickup is in the past.';
 	public const SLOT_COUNT_OUT_OF_RANGE = 'Slot value out of range.';
 	public const MORE_OCCUPIED_SLOTS = 'Slot value is smaller then occuied slots.';
 	public const INVALID_STORE = 'Store does not exist.';
 
-	public function __construct(string $message = '', int $code = 0, Throwable $previous = null)
+	public function __construct(string $message = '', int $code = 0)
 	{
-		parent::__construct($message, $code, $previous);
+		parent::__construct($message, $code);
 	}
 
 	public function __toString(): string